Moved getAdminPagePropsFromPath function to utils file and removed small part of it to make it slightly more general. This mostly to reduce number of lines in Admin component.
