+++ client/src/components/admin/utils.ts
@@ -22,3 +22,32 @@ export function getGETParamsStringFromObject(getParamsObj?: NodeJS.Dict<string|n
         return '';
     }
 }
+
+/**
+ * Function that parses the search string of the path and GET-parameters to desired format. You probably want to pass in
+ * window.location.search as the searchString argument.
+ * TODO Probably possible to use some inbuilt functions to do this with fewer steps
+ * @param searchString The get parameters given in path by the search parameter.
+ * @return Returns a dict representing the values in the get queries.
+ */
+export function GETParamsToDict(searchString: string): NodeJS.Dict<string | undefined> {
+    // Remove leading ?
+    if (searchString.length > 0 && searchString[0] === '?') {
+        searchString = searchString.substring(1);
+    }
+    // Split over different args
+    const splitArgs = searchString.split('&');
+
+    // For each element, if it has value, use that. Else just set the key with undefined.
+    const getParamsReturn: NodeJS.Dict<string | undefined> = {};
+    splitArgs.forEach((item) => {
+        if (item.includes('=')) {
+            const [key, val] = item.split('=', 1);
+            getParamsReturn[key] = val;
+        } else if (item !== '') {
+            getParamsReturn[item] = undefined;
+        }
+    });
+
+    return getParamsReturn;
+}
